# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CourseCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Title",
            "type": "string"
          },
          "visibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Visibility"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "title"
        ],
        "title": "CourseCreate",
        "type": "object"
      },
      "CourseDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lecture_count": {
            "default": 0,
            "title": "Lecture Count",
            "type": "integer"
          },
          "lectures": {
            "items": {
              "$ref": "#/components/schemas/LectureOut"
            },
            "title": "Lectures",
            "type": "array"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "visibility",
          "owner_id",
          "created_at",
          "updated_at",
          "lectures"
        ],
        "title": "CourseDetail",
        "type": "object"
      },
      "CourseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lecture_count": {
            "default": 0,
            "title": "Lecture Count",
            "type": "integer"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "visibility",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LectureCreate": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "content_type": {
            "examples": [
              "video/mp4"
            ],
            "pattern": "^video/[\\w.+-]+$",
            "title": "Content Type",
            "type": "string"
          },
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "filename": {
            "examples": [
              "6.006-lecture-01.mp4"
            ],
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "licence": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "CC BY-NC-SA 4.0"
            ],
            "title": "Licence"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "filename",
          "content_type"
        ],
        "title": "LectureCreate",
        "type": "object"
      },
      "LectureCreated": {
        "properties": {
          "lecture": {
            "$ref": "#/components/schemas/LectureOut"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadTarget"
          }
        },
        "required": [
          "lecture",
          "upload"
        ],
        "title": "LectureCreated",
        "type": "object"
      },
      "LectureOut": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "licence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licence"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/LectureStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        },
        "required": [
          "id",
          "title",
          "course_id",
          "status",
          "source_filename",
          "content_type",
          "size_bytes",
          "duration_s",
          "content_hash",
          "licence",
          "attribution",
          "visibility",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "LectureOut",
        "type": "object"
      },
      "LectureStatus": {
        "enum": [
          "awaiting_upload",
          "uploaded",
          "processing",
          "ready",
          "failed"
        ],
        "title": "LectureStatus",
        "type": "string"
      },
      "LectureUpdate": {
        "description": "Only the fields sent are changed. `course_id: null` takes a lecture out of its course.\nOnly admins change `visibility`.",
        "properties": {
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "visibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Visibility"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "LectureUpdate",
        "type": "object"
      },
...
      "MediaOut": {
        "description": "Where the browser plays the lecture from: a presigned URL straight to storage.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_in_s": {
            "title": "Expires In S",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "content_type",
          "expires_in_s"
        ],
        "title": "MediaOut",
        "type": "object"
      },
...
      "UploadTarget": {
        "description": "Send the file with `method` to `url`, including `headers` exactly as given.",
        "properties": {
          "expires_in_s": {
            "title": "Expires In S",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "method": {
            "const": "PUT",
            "default": "PUT",
            "title": "Method",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "headers",
          "expires_in_s"
        ],
        "title": "UploadTarget",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Visibility": {
        "enum": [
          "public",
          "private"
        ],
        "title": "Visibility",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Lecture Summariser API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/courses": {
      "get": {
        "operationId": "list_courses_v1_courses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseOut"
                  },
                  "title": "Response List Courses V1 Courses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Courses",
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "create_course_v1_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/v1/courses/{course_id}": {
      "delete": {
        "operationId": "delete_course_v1_courses__course_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "get_course_v1_courses__course_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/v1/lectures": {
      "get": {
        "description": "The lectures the caller may read: the public ones, and their own.",
        "operationId": "list_lectures_v1_lectures_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LectureOut"
                  },
                  "title": "Response List Lectures V1 Lectures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lectures",
        "tags": [
          "lectures"
        ]
      },
      "post": {
        "description": "A signed-in user's lecture is private to them. Without sign-in (local), it's public.",
        "operationId": "create_lecture_v1_lectures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LectureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Lecture",
        "tags": [
          "lectures"
        ]
      }
    },
    "/v1/lectures/{lecture_id}": {
      "get": {
        "operationId": "get_lecture_v1_lectures__lecture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lecture",
        "tags": [
          "lectures"
        ]
      },
      "patch": {
        "description": "Change only the fields sent: the course (one of the caller's own), or, for admins, who\ncan see the lecture.",
        "operationId": "update_lecture_v1_lectures__lecture_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LectureUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Lecture",
        "tags": [
          "lectures"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/complete-upload": {
      "post": {
        "description": "Idempotent: confirming an upload that's already confirmed returns the lecture unchanged.",
        "operationId": "complete_upload_v1_lectures__lecture_id__complete_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Upload",
        "tags": [
          "lectures"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/media": {
      "get": {
        "description": "A presigned URL for playing the uploaded video. Browsers play MP4 (H.264/AAC) directly;\nother formats will need the HLS renditions planned for later.",
        "operationId": "media_v1_lectures__lecture_id__media_get",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Media",
        "tags": [
          "lectures"
        ]
      }
    },
...
  }
}